fix(v2-p5): OIDC public path routing — match discovery doc
/qa 在 V2-P6 sessions 期間掃出 OIDC discovery doc 公告 4 endpoint 中
2 個 routing 錯（authorization_endpoint, token_endpoint, revocation_endpoint
不指對地方）。Pre-existing 自 V2-P1 (#256)。

Root cause: V2-P1 OAuth Client（去 Google）佔了 `/api/oauth/authorize`，
V2-P4 OAuth Server 只能用 `/server-authorize`。Discovery doc 然後公告
`/authorize`（公開 OIDC 慣例），於是 external client 跟 doc 走會打錯路。

Fix（autoplan finding 4 推薦的 freeze /oauth/* 為 public）：

File renames：
- server-authorize.ts → authorize.ts （public OAuth Server endpoint）
- server-token.ts → token.ts
- server-consent.ts → consent.ts
- 舊 authorize.ts (Google client) → login/google.ts （internal subdir）
- callback.ts (Google callback) → callback/google.ts （symmetric subdir）

新增：
- revoke.ts — RFC 7009 token revocation endpoint（discovery 公告但之前缺）
  - silent 200 on unknown token (anti-scanning per §2.2)
  - silent 200 on wrong client (anti-leak)
  - revoke refresh_token cascades to all access_tokens via grantId

Internal redirects 跟 frontend 連結同步：
- consent.ts → /api/oauth/authorize （原 server-authorize）
- LoginPage.tsx href: /api/oauth/login/google （原 authorize?provider=google）
- ConsentPage.tsx redirect: /api/oauth/authorize

Tests：6 file renames + URL string updates 一致化、新增 oauth-revoke.test.ts
（6 cases — RFC 7009 silent 200 / cross-client / token_type_hint 等）。

Drop redundant `?provider=google` validation：path now explicit。

Discovery doc 的 endpoint 路徑沒改 — 本來就指 `/authorize`、`/token`、`/revoke`。
重點是後端終於對齊。

Verified locally:
  curl /.well-known/openid-configuration → all 5 endpoints respond:
    authorization_endpoint /api/oauth/authorize  → OAuth Server (200/302)
    token_endpoint /api/oauth/token              → 400 missing grant_type ✓
    userinfo_endpoint /api/oauth/userinfo        → 401 (auth required) ✓
    revocation_endpoint /api/oauth/revoke        → 400 missing token ✓
    jwks_uri /api/oauth/.well-known/jwks.json    → 200 ✓

894/894 vitest 全綠 + tsc clean.

⚠️ DEPLOY NOTE：
1. Google Cloud Console redirect_uri 改 `/api/oauth/callback` → `/api/oauth/callback/google`
2. 任何 internal monitoring / runbook 提到 server-authorize / server-token / server-consent 路徑需更新
